Fixes based on code review
