+++ lib/src/fetch/config_fetcher.dart
@@ -1,15 +1,13 @@
 import 'dart:async';
-import 'dart:convert';
 import 'package:dio/dio.dart';
 
+import '../Utils.dart';
 import '../error_reporter.dart';
 import '../entry.dart';
 import '../data_governance.dart';
 import '../configcat_options.dart';
 import '../constants.dart';
 import '../json/config.dart';
-import '../json/segment.dart';
-import '../json/setting.dart';
 import '../log/configcat_logger.dart';
 
 enum _Status { fetched, notModified, failure }
@@ -182,7 +180,7 @@ class ConfigFetcher implements Fetcher {
         var configJson = response.data.toString();
         Config config;
         try {
-          config = _deserializeConfig(configJson);
+          config = Utils.deserializeConfig(configJson);
         } catch (e) {
           String error =
               "Fetching config JSON was successful but the HTTP response content was invalid.";
@@ -228,21 +226,5 @@ class ConfigFetcher implements Fetcher {
           s);
       return FetchResponse.failure(e.toString(), true);
     }
-  }
-
-  static Config _deserializeConfig(String configJson) {
-    final decoded = jsonDecode(configJson);
-    Config config = Config.fromJson(decoded);
-    String salt = config.preferences.salt;
-    if (salt.isEmpty) {
-      throw ArgumentError("Config JSON salt is missing.");
-    }
-    List<Segment> segments = config.segments;
-
-    for (Setting setting in config.entries.values) {
-      setting.salt = salt;
-      setting.segments = segments;
-    }
-    return config;
   }
 }
